5310: Heroku cron jobs on separate worker
