+++ Procfile
@@ -1,0 +1,2 @@
+web: yarn start
+worker: node dist/server/worker
