+++ package.json
@@ -25,6 +25,8 @@
     "run:dev:server": "yarn migration-public:run && ts-node-dev --files --exit-child src/server/start",
     "run:prod": "run-s clean build start",
     "run:prod:server": "node dist/server/start",
+    "run:heroku:start": "heroku local:start",
+    "run:heroku:local": "run-s clean build run:heroku:start",
     "start": "run-s migration-public:run run:prod:server",
     "//test": "mochapack --webpack-config webpack.config.babel.test.js",
     "test": "run-p test:unit test:integration",
@@ -234,6 +236,7 @@
     "marked": "^4.0.10",
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
+    "node-cron": "^4.2.1",
     "node-schedule": "^2.0.0",
     "nodemailer": "^6.7.3",
     "npm-run-all": "^4.1.5",
